Use MS devcontainer base, docker cp R/Python libs
Base on MS devcontainer Jammy image to match OpenSAFELY base image.

Install R/Python and their dependencies using the respective
requirements.txt from the `main` branch on github.
This file is not copied to the built docker action images so we cannot
copy it from their containers.

Use `docker cp` to copy the R/Python library files to the host filesystem
which avoids any of the previously seen issues with mounting a running
container.

However, these library directories don't contain any symlinks
(verified with `find . -type l`) so one of the major issues with
mounting is not of concern here so commented-out mount-based
alternative is provided here which has better performance than the
`docker cp` approach.